+++ ChangeLog.md
@@ -1,6 +1,7 @@
 #2.2.12
   * Tidy up old references to coriolis.io
   * Add ability to add and remove special effects to weapon modifications
+  * Add weapon engineering information to Damage Dealt section
 
 #2.2.11
   * Add help system and initial help file